requirements.txt
gooey==1.0.8.1
pandas==2.0.3
mysql-connector-python==8.0.33
numpy==1.24.3